+++ packages/backend/src/engine/branch-linker/lambda/branch_linker_types.ts
@@ -1,0 +1,4 @@
+export type AmplifyBranchLinkerCustomResourceProps = {
+  backendId: string;
+  branchName: string;
+};