+++ packages/backend-auth/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @aws-amplify/backend-auth
 
+## 0.2.2
+
+### Patch Changes
+
+- Updated dependencies [457b1662]
+  - @aws-amplify/plugin-types@0.3.0
+  - @aws-amplify/auth-construct-alpha@0.2.2
+
 ## 0.2.1
 
 ### Patch Changes