+++ packages/backend-function/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @aws-amplify/backend-function
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies [457b1662]
+  - @aws-amplify/plugin-types@0.3.0
+
 ## 0.1.1
 
 ### Patch Changes