+++ packages/backend-secret/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @aws-amplify/backend-secret
 
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies [457b1662]
+  - @aws-amplify/plugin-types@0.3.0
+  - @aws-amplify/platform-core@0.1.2
+
 ## 0.2.0
 
 ### Minor Changes